+++ tests/testthat/helper-objects.R
@@ -3,6 +3,7 @@ assign("message_variable_importance", value = TRUE, envir = DALEX:::.DALEX.env)
 assign("message_prediction_breakdown", value = TRUE, envir = DALEX:::.DALEX.env)
 assign("message_partial_dependency", value = TRUE, envir = DALEX:::.DALEX.env)
 assign("message_accumulated_dependency", value = TRUE, envir = DALEX:::.DALEX.env)
+assign("message_variable_effect", value = TRUE, envir = DALEX:::.DALEX.env)
 
 
 library("DALEX")